Removed dependencies cycles and set up import/no-cycle to be reported as error
